Move `resource` struct to be platform-agnostic
