+++ aws/plugin/cloudfront/cloudfront.go
@@ -11,7 +11,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/cloudfront/types"
 
 	awsconnector "github.com/magneticstain/ip-2-cloudresource/aws/aws_connector"
-	generalResource "github.com/magneticstain/ip-2-cloudresource/aws/resource"
+	generalResource "github.com/magneticstain/ip-2-cloudresource/resource"
 	"github.com/magneticstain/ip-2-cloudresource/utils"
 )
